Simple conversion of Emit's functions to uint, and added the uint version of LLVM functions that were wrapped in CodeGen

+++ src/Syntax.hs
@@ -33,7 +33,7 @@ reservedWords =
   ]
 
 data Expr
-  = NumberExp Double
+  = NumberExp Integer
   | EmptyExp
   | BoolExp Bool
   | VarExp String
